refactor: make validation logic reusable
